ruby: object proxies now call methods correctly.
Fixed a bug in method missing such that it uses the stringified version of the method name, not the symbol, when passing to the other side.

Fix a bug in proxy overrides to Object which caused it to use the wrong variable name to pass args.

Ensure that nil transfers as a primitive, not a proxy.

Actually pass the real (non-proxy) class name through when calling object methods ...though this is really only used for getting special opts which is not implemented yet for ruby.

Add RMI_FLOP environment variable to cause a server to not pass back an exception but throw it in place.

Fix a bug in which a forked-pipes client didn't know the pid of its server.

Updated debugging code before responding to a request

Updated test 01 to execute more of its original logic.